# backend/swagger-output.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Service Tracking Platform API",
        "description": "REST API for the Service Tracking Platform with AI powered service request classification and technician assignment.\n        \n        Features:\n        - JWT-based authentication and authorization\n        - AI Smart Request Creation\n        - AI Technician Recommendation\n        - Status Transition Management\n        - Real-time Assignment Workflow\n        ",
        "contact": {
            "name": "Mileta Lvova",
            "email": "mileta279@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/services/": {
            "get": {
                "tags": ["Services"],
                "summary": "Get all services",
                "description": "Endpoint to get all services",
                "responses": {
                    "200": {
                        "description": "List of services retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List of services"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Service"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError",
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            },
            "post": {
                "tags": ["Services"],
                "summary": "Creates a new service",
                "description": "Endpoint to create a new service",
                "responses": {
                    "201": {
                        "description": "Service created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "number",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service created"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Service"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError",
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ServiceInput"
                            }
                        }
                    }
                }
            }
        },
        "/api/services/{id}": {
            "get": {
                "tags": ["Services"],
                "summary": "Get service by ID",
                "description": "Endpoint to get details of a specific service by its ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "example": "integer"
                                },
                                "example": {
                                    "type": "number",
                                    "example": 1
                                }
                            }
                        },
                        "description": "Service ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of a service retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service details"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Service"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "description": "Not Found",
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError",
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            },
            "put": {
                "tags": ["Services"],
                "summary": "Updates a service",
                "description": "Endpoint to update a service",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "example": "integer"
                                },
                                "example": {
                                    "type": "number",
                                    "example": 1
                                }
                            }
                        },
                        "description": "Service ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service updated"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Service"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden",
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound",
                        "description": "Not Found"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError",
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ServiceUpdateInput"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Services"],
                "summary": "Delete a service",
                "description": "Endpoint to delete a service",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "example": "integer"
                                },
                                "example": {
                                    "type": "number",
                                    "example": 1
                                }
                            }
                        },
                        "description": "Service ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service deleted"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden",
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound",
                        "description": "Not Found"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError",
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "JWT": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Authorization header using the Bearer scheme. \n                Enter ONLY your JWT token in the input below (without 'Bearer ' prefix).\n                \n                Example: \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\""
            }
        },
        "schemas": {
...
            "Service": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "object"
                    },
                    "properties": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "integer"
                                    },
                                    "example": {
                                        "type": "number",
                                        "example": 1
                                    }
                                }
                            },
                            "specialization": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "Plumbing"
                                    }
                                }
                            },
                            "description": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "All plumbing related services"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "ServiceInput": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "object"
                    },
                    "properties": {
                        "type": "object",
                        "properties": {
                            "specialization": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "Plumbing"
                                    }
                                }
                            },
                            "description": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "All plumbing related services"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "ServiceUpdateInput": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "object"
                    },
                    "properties": {
                        "type": "object",
                        "properties": {
                            "specialization": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "Plumbing"
                                    }
                                }
                            },
                            "description": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "string"
                                    },
                                    "example": {
                                        "type": "string",
                                        "example": "All plumbing related services"
                                    }
                                }
                            }
                        }
                    }
                }
            },
...
        },
        "responses": {
            "BadRequest": {
                "description": "Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Unauthorized - Authentication required",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Forbidden - Insufficient permissions",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            },
            "InternalServerError": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ApiError"
                        }
                    }
                }
            }
        }
    },
    "security": [
        {
            "JWT": []
        }
    ]
}